service_b/requirements.txt
fastapi==0.103.2
uvicorn==0.24.0
grpcio==1.59.0
grpcio-tools==1.59.0
httpx==0.25.0
pydantic==1.10.12
